General: Update translation_authors.yaml
I was one of the translators for the GAMEBOY portion, and assisted in translating 60-70% of the content in the beginning. Then the translation program was put on hold due to exams, and then Carryspider completed the rest of the section.
Add my name as one of the contributors to this branch. However, I would like to thank Carryspider for his contribution, without which the article would not have been completed as quickly as it was.